Fix percentage calc, move connectionHandlerThreadPoolSize param

+++ src/TestClient/TestClient.java
@@ -14,11 +14,11 @@ public class TestClient {
             byte[] buf = new byte[Const.bufferSize];
             int read = 0;
             long readTotal = 0;
-            int len = fis.available();
+            long len = fis.available();
             while((read = fis.read(buf, 0, buf.length)) != -1) {
                 readTotal += read;
                 os.write(buf, 0, read);
-                int perc = (int)((float)readTotal/len * 100);
+                int perc = (int)((double)(readTotal/len) * 100);
                 System.out.print("\r" +  perc + "% written");
             }
             System.out.println();
